récupération d'information et connexion
